# backendRL/src/main/java/org/example/backendrl/config/api-docs.yaml
openapi: 3.1.0
info:
  title: RecordingLogs API
  description: "Music rating and recommendation platform: genres, artists, albums,\
    \ songs, ratings, reviews, album lists, favorites and follows."
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /api/artists:
    get:
      tags:
      - Artists
      summary: Get all artists
      operationId: getAllArtists
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: List of artists
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArtistResponse"
    post:
      tags:
      - Artists
      summary: Create an artist
      operationId: createArtist
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtistRequest"
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Invalid request body
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "201":
          description: Artist created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ArtistResponse"
...
  /api/albums:
    get:
      tags:
      - Albums
      summary: Get all albums
      operationId: getAllAlbums
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: List of albums
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlbumResponse"
    post:
      tags:
      - Albums
      summary: Create an album
      operationId: createAlbum
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlbumRequest"
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Invalid request body
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "201":
          description: Album created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AlbumResponse"
  /api/albums/{albumId}/songs/{songId}:
    post:
      tags:
      - Albums
      summary: Add a song to an album's tracklist
      operationId: addSongToAlbum
      parameters:
      - name: albumId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: songId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlbumSongRequest"
        required: true
      responses:
        "404":
          description: Album or song not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Song is already on this album
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Invalid request body
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "201":
          description: Song added to album
    delete:
      tags:
      - Albums
      summary: Remove a song from an album's tracklist
      operationId: removeSongFromAlbum
      parameters:
      - name: albumId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: songId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: "Album, song or link not found"
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "204":
          description: Song removed from album
...
  /api/songs/{songId}/albums:
    get:
      tags:
      - Songs
      summary: Get all albums containing a song
      operationId: getAlbumsForSong
      parameters:
      - name: songId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Song not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: List of albums
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlbumSummaryResponse"
...
  /api/artists/{artistId}/albums:
    get:
      tags:
      - Artists
      summary: Get all albums of an artist
      operationId: getAlbumsForArtist
      parameters:
      - name: artistId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Artist not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: List of albums
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlbumSummaryResponse"
  /api/albums/{albumId}/songs:
    get:
      tags:
      - Albums
      summary: Get the tracklist of an album
      operationId: getSongsForAlbum
      parameters:
      - name: albumId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Album not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: List of songs
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlbumSongResponse"
...
  /api/albums/{albumId}/artists:
    get:
      tags:
      - Albums
      summary: Get all artists of an album
      operationId: getArtistsForAlbum
      parameters:
      - name: albumId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Album not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: List of artists
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArtistSummaryResponse"
components:
  schemas:
...
    ArtistRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        biography:
          type: string
      required:
      - name
    ArtistResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        biography:
          type: string
    AlbumRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        releaseDate:
          type: string
          format: date
        coverUrl:
          type: string
      required:
      - title
    AlbumResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        releaseDate:
          type: string
          format: date
        coverUrl:
          type: string
...
    AlbumSongRequest:
      type: object
      properties:
        trackNumber:
          type: integer
          format: int32
          minimum: 1
        discNumber:
          type: integer
          format: int32
          minimum: 1
      required:
      - trackNumber
...
    ArtistSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    AlbumSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        releaseDate:
          type: string
          format: date
        coverUrl:
          type: string
...
    AlbumSongResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        albumId:
          type: integer
          format: int64
        songId:
          type: integer
          format: int64
        songTitle:
          type: string
        trackNumber:
          type: integer
          format: int32
        discNumber:
          type: integer
          format: int32